test: 입/출력 기능 분리 테스트 코드 작성
