+++ src/test/java/io/InputHandlerTest.java
@@ -33,7 +33,7 @@ class InputHandlerTest {
     }
 
     @Test
-    void 시도할_횟수_입력하기() {
+    void 시도할_횟수_입력_부분_확인하기() {
         // given
         Console.close();
         String simulatedInput = "4";
@@ -41,23 +41,9 @@ class InputHandlerTest {
         System.setIn(in);
 
         // when
-        int inputRacingCount = inputHandler.getInputRacingCount();
+        String inputRacingCount = inputHandler.getInputRacingCount();
 
         // then
-        assertThat(inputRacingCount).isEqualTo(Integer.parseInt(simulatedInput.trim()));
-    }
-
-    @Test
-    void 시도할_횟수_양수가_아닌_경우_에러_던지기() {
-        // given
-        Console.close();
-        String simulatedInput = "-1";
-        ByteArrayInputStream in = new ByteArrayInputStream(simulatedInput.getBytes());
-        System.setIn(in);
-
-        // when // then
-        assertThatThrownBy(() -> inputHandler.getInputRacingCount())
-                .isInstanceOf(IllegalArgumentException.class)
-                .hasMessage("시도할 횟수는 양수 이여야합니다.");
+        assertThat(inputRacingCount).isEqualTo(simulatedInput);
     }
 }
